requirements.txt
# requirements.txt

aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
cachelib==0.2.0
certifi==2025.4.26
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mezmorize==0.28.2
multidict==6.4.3
propcache==0.3.1
py-cord==2.6.1
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
typing_extensions==4.13.2
Werkzeug==2.0.0
yarl==1.20.0
